Fix unexpanded %s in leftover values error messages
Before:

        myfunction: '%s' parameter still has values that haven't been checked.

Now, %s is expanded properly:

        myfunction: 'num_samples' parameter still has values that haven't been checked.

Parameter `error_message` of functions `check_for_leftover_values` and
`check_for_leftover_values_list` isn't actually the error message but
rather a printf-style format string (containing one `%s`) used to build
the error message. That means that it must be the first argument of
`cm_print_error`. The previous version used another `%s` to output the
contents of `error_message`, so the format string was printed but not
expanded.

Renamed the parameter to `error_message_fmt` for emphasis.

Reviewed-by: Andreas Schneider <asn@cryptomilk.org>