apps/testing/cmocka/cmocka: Fix bug VELAPLATFO-5208, show skipped case msg in log.
When run cmocka test, if has skip filter defined in test command, the number of skipped test case isn't increased.
In this change, if has skip filter in test command, set the status of the skipped case as CM_TEST_SKIPPED, then the number of skipped test case will increase,and then the skipped msg will be printed in log.